backend/ingredients/views.py
from django_filters.rest_framework import DjangoFilterBackend
from rest_framework import viewsets

from recipes.filters import IngredientFilter

from .models import Ingredient
from .serializers import IngredientSerializer


class IngredientViewSet(viewsets.ReadOnlyModelViewSet):
    """Вьюсет для работы с ингредиентами."""
    queryset = Ingredient.objects.all()
    serializer_class = IngredientSerializer
    filter_backends = (DjangoFilterBackend, )
    filterset_class = IngredientFilter
